Update Droid-Toolbox.ino
fixed location beacon names; they were not correct. we can argue about group 5. should it be called oga's. should it be called droid depot? all stock chips have a single noise in that group and it's always an alarm noise. so i'm changing the name to Alert. (should it be Alarm?)